refactor(viem)!: remove custom L1 actions in favor of plain viem contract actions
Remove the addInput, deposit*, executeOutput, validateOutput and
estimate*Gas actions and the walletActionsL1 / publicActionsL1
decorators. These were thin wrappers around viem's writeContract /
readContract / estimateContractGas with the generated ABIs, which are
already exported through the @cartesi/viem/abi entrypoint. Wagmi users
get the same functionality from the generated actions and hooks in
@cartesi/wagmi, or can generate their own code with
@cartesi/wagmi-plugin.

Export the toEVM helper (and its ExecuteOutputArgs type), which
converts an Output returned by the node API into the arguments of
IApplication.executeOutput / validateOutput.

Replace the per-action documentation pages with a single L1 Contract
Interactions guide, and update the README, examples and snippets to
plain viem usage.
