--- apps/docs/pages/viem/addInput.mdx
@@ -1,42 +1,0 @@
-# addInput
-
-## Usage
-
-```ts twoslash
-import { http } from "viem";
-import { createWalletClient } from "viem";
-import { createCartesiPublicClient, walletActionsL1 } from "@cartesi/viem";
-import { privateKeyToAccount } from "viem/accounts";
-import { mainnet } from "viem/chains";
-
-const chain = mainnet;
-const account = privateKeyToAccount("0x...");
-const walletClient = createWalletClient({ chain, transport: http() }).extend(walletActionsL1());
-
-const hash = await walletClient.addInput({
-  account,
-  application: "0x...",
-  chain,
-  payload: "0x...",
-});
-```
-
-## Parameters
-
-```ts twoslash
-import type { AddInputParameters } from "@cartesi/viem";
-// {
-//   account: Account | Address;
-//   application: Address;
-//   chain?: Chain;
-//   payload: Hex;
-//   gas?: bigint | null;
-//   ...other viem FormattedTransactionRequest fields
-// }
-```
-
-## Return Type
-
-```ts twoslash
-// Hash (string)
-```
