--- apps/docs/pages/viem/depositERC721Token.mdx
@@ -1,48 +1,0 @@
-# depositERC721Token
-
-## Usage
-
-```ts twoslash
-import { http } from "viem";
-import { createWalletClient } from "viem";
-import { createCartesiPublicClient, walletActionsL1 } from "@cartesi/viem";
-import { privateKeyToAccount } from "viem/accounts";
-import { mainnet } from "viem/chains";
-
-const chain = mainnet;
-const account = privateKeyToAccount("0x...");
-const walletClient = createWalletClient({ chain, transport: http() }).extend(walletActionsL1());
-
-const hash = await walletClient.depositERC721Token({
-  account,
-  application: "0x...",
-  chain,
-  token: "0x...",
-  tokenId: 1n,
-  baseLayerData: "0x",
-  execLayerData: "0x",
-});
-```
-
-## Parameters
-
-```ts twoslash
-import type { DepositERC721TokenParameters } from "@cartesi/viem";
-// {
-//   account: Account | Address;
-//   application: Address;
-//   chain?: Chain;
-//   token: Address;
-//   tokenId: bigint;
-//   baseLayerData: Hex;
-//   execLayerData: Hex;
-//   gas?: bigint | null;
-//   ...other viem FormattedTransactionRequest fields
-// }
-```
-
-## Return Type
-
-```ts twoslash
-// Hash (string)
-```
