--- apps/docs/pages/viem/depositEther.mdx
@@ -1,44 +1,0 @@
-# depositEther
-
-## Usage
-
-```ts twoslash
-import { http, parseUnits } from "viem";
-import { createWalletClient } from "viem";
-import { createCartesiPublicClient, walletActionsL1 } from "@cartesi/viem";
-import { privateKeyToAccount } from "viem/accounts";
-import { mainnet } from "viem/chains";
-
-const chain = mainnet;
-const account = privateKeyToAccount("0x...");
-const walletClient = createWalletClient({ chain, transport: http() }).extend(walletActionsL1());
-
-const hash = await walletClient.depositEther({
-  account,
-  application: "0x...",
-  chain,
-  execLayerData: "0x",
-  value: parseUnits("1", 18),
-});
-```
-
-## Parameters
-
-```ts twoslash
-import type { DepositEtherParameters } from "@cartesi/viem";
-// {
-//   account: Account | Address;
-//   application: Address;
-//   chain?: Chain;
-//   execLayerData: Hex;
-//   value: bigint;
-//   gas?: bigint | null;
-//   ...other viem FormattedTransactionRequest fields
-// }
-```
-
-## Return Type
-
-```ts twoslash
-// Hash (string)
-```
