--- apps/docs/pages/viem/depositSingleERC1155Token.mdx
@@ -1,50 +1,0 @@
-# depositSingleERC1155Token
-
-## Usage
-
-```ts twoslash
-import { http } from "viem";
-import { createWalletClient } from "viem";
-import { createCartesiPublicClient, walletActionsL1 } from "@cartesi/viem";
-import { privateKeyToAccount } from "viem/accounts";
-import { mainnet } from "viem/chains";
-
-const chain = mainnet;
-const account = privateKeyToAccount("0x...");
-const walletClient = createWalletClient({ chain, transport: http() }).extend(walletActionsL1());
-
-const hash = await walletClient.depositSingleERC1155Token({
-  account,
-  application: "0x...",
-  chain,
-  token: "0x...",
-  tokenId: 1n,
-  value: 1n,
-  baseLayerData: "0x",
-  execLayerData: "0x",
-});
-```
-
-## Parameters
-
-```ts twoslash
-import type { DepositSingleERC1155TokenParameters } from "@cartesi/viem";
-// {
-//   account: Account | Address;
-//   application: Address;
-//   chain?: Chain;
-//   token: Address;
-//   tokenId: bigint;
-//   value: bigint;
-//   baseLayerData: Hex;
-//   execLayerData: Hex;
-//   gas?: bigint | null;
-//   ...other viem FormattedTransactionRequest fields
-// }
-```
-
-## Return Type
-
-```ts twoslash
-// Hash (string)
-```
