--- apps/docs/pages/viem/estimateAddInputGas.mdx
@@ -1,5 +1,0 @@
-# estimateAddInputGas
-
-:::note
-TODO
-:::
