--- apps/docs/pages/viem/estimateDepositBatchERC1155TokenGas.mdx
@@ -1,5 +1,0 @@
-# estimateDepositBatchERC1155TokenGas
-
-:::note
-TODO
-:::
