--- apps/docs/pages/viem/estimateDepositERC20TokenGas.mdx
@@ -1,5 +1,0 @@
-# estimateDepositERC20TokensGas
-
-:::note
-TODO
-:::
