--- apps/docs/pages/viem/estimateDepositERC721TokenGas.mdx
@@ -1,5 +1,0 @@
-# estimateDepositERC721TokenGas
-
-:::note
-TODO
-:::
