--- apps/docs/pages/viem/estimateDepositEtherGas.mdx
@@ -1,5 +1,0 @@
-# estimateDepositEtherGas
-
-:::note
-TODO
-:::
