--- apps/docs/pages/viem/estimateDepositSingleERC1155TokenGas.mdx
@@ -1,5 +1,0 @@
-# estimateDepositSingleERC1155TokenGas
-
-:::note
-TODO
-:::
