--- apps/docs/pages/viem/estimateExecuteOutputGas.mdx
@@ -1,5 +1,0 @@
-# estimateExecuteOutputGas
-
-:::note
-TODO
-:::
