--- apps/docs/pages/viem/executeOutput.mdx
@@ -1,42 +1,0 @@
-# executeOutput
-
-## Usage
-
-```ts
-import { http } from "viem";
-import { createWalletClient } from "viem";
-import { createCartesiPublicClient, walletActionsL1 } from "@cartesi/viem";
-import { privateKeyToAccount } from "viem/accounts";
-import { mainnet } from "viem/chains";
-
-const chain = mainnet;
-const account = privateKeyToAccount("0x...");
-const walletClient = createWalletClient({ chain, transport: http() }).extend(walletActionsL1());
-
-const hash = await walletClient.executeOutput({
-  account,
-  application: "0x...",
-  chain,
-  output: {/* Output object */},
-});
-```
-
-## Parameters
-
-```ts twoslash
-import type { ExecuteOutputParameters } from "@cartesi/viem";
-// {
-//   account: Account | Address;
-//   application: Address;
-//   chain?: Chain;
-//   output: Output;
-//   gas?: bigint | null;
-//   ...other viem FormattedTransactionRequest fields
-// }
-```
-
-## Return Type
-
-```ts twoslash
-// Hash (string)
-```
