+++ apps/docs/pages/viem/index.mdx
@@ -2,24 +2,12 @@
 
 This is a [viem](https://viem.sh) extension library that helps with the development of Cartesi applications, especially frontend applications that run in the browser.
 
-It provides actions that complement the L1 public and wallet actions, by providing convenient methods for interacting with Cartesi rollups [smart contracts](https://github.com/cartesi/rollups-contracts) infrastructure.
+It provides L2 public actions and a new `createCartesiPublicClient` client constructor that allows to interact with the Cartesi rollups [JSON-RPC API](https://playground.open-rpc.org/?schemaUrl=https://raw.githubusercontent.com/cartesi/rollups-node/refs/tags/v2.0.0-alpha.10/internal/jsonrpc/jsonrpc-discover.json&uiSchema[appBar][ui:title]=Cartesi&uiSchema[appBar][ui:logoUrl]=https://cartesi.io/favicon.svg&uiSchema[appBar][ui:transports]=false&uiSchema[appBar][ui:input]=false&uiSchema[appBar][ui:edit]=false&uiSchema[appBar][ui:examplesDropdown]=false&uiSchema[appBar][ui:splitView]=false).
 
-It also provides L2 public actions and a new `createCartesiPublicClient` client constructor that allows to interact with the Cartesi rollups [JSON-RPC API](https://playground.open-rpc.org/?schemaUrl=https://raw.githubusercontent.com/cartesi/rollups-node/refs/tags/v2.0.0-alpha.10/internal/jsonrpc/jsonrpc-discover.json&uiSchema[appBar][ui:title]=Cartesi&uiSchema[appBar][ui:logoUrl]=https://cartesi.io/favicon.svg&uiSchema[appBar][ui:transports]=false&uiSchema[appBar][ui:input]=false&uiSchema[appBar][ui:edit]=false&uiSchema[appBar][ui:examplesDropdown]=false&uiSchema[appBar][ui:splitView]=false).
+## L1 Contract Interactions
 
-## Wallet L1 Actions
-
-Provides convenient methods to interact with Cartesi rollups smart contracts.
-The `addInput` allows to send generic inputs to the `InputBox`, and there are methods to deposit assets to each of the asset portals: ERC-20, ERC-721 and ERC-1155.
-There is also a method to execute outputs (vouchers).
-
-If the user wants to use directly viem methods, or wagmi hooks, they can also use the `@cartesi/viem/abi` package to import properly type ABIs for all contracts managed by Cartesi rollups.
-
-## Public L1 Actions
-
-Provides convenient methods to interact with view methods of Cartesi rollups smart contracts.
-This includes gas estimation methods for all wallet L1 actions, and also the `validateOutput` method.
-
-If the user wants to use directly viem methods, or wagmi hooks, they can also use the `@cartesi/viem/abi` package to import properly type ABIs for all contracts managed by Cartesi rollups.
+Interactions with the Cartesi rollups [smart contracts](https://github.com/cartesi/rollups-contracts) infrastructure, like sending inputs, depositing assets and executing outputs, are done with viem's own contract actions, using the properly typed ABIs and deployment addresses exported by the `@cartesi/viem/abi` entrypoint.
+Check the [L1 Contract Interactions](/viem/l1) guide for details.
 
 ## Public L2 Actions
