+++ packages/viem/README.md
@@ -2,33 +2,7 @@
 
 Viem provides an [extension mechanism to clients](https://viem.sh/docs/clients/custom) that can provide a convenient and familiar API for application developers based on L2 solutions.
 
-This extension provides extensions for L1 operations and L2 operations related to Cartesi.
-
-## WalletClient L1
-
-Transactions related to Cartesi applications are sent directly to the base layer (L1).
-This extension provides convenient methods for [sending inputs](https://docs.cartesi.io/cartesi-rollups/1.5/rollups-apis/json-rpc/input-box/#addinput) and interacting with the [portals](https://docs.cartesi.io/cartesi-rollups/1.5/rollups-apis/json-rpc/portals/ERC20Portal/) provided by Cartesi Rollups.
-
-- addInput
-- depositEther
-- depositERC20Tokens
-- depositERC721Token
-- depositSingleERC1155Token
-- depositBatchERC1155Token
-- executeOutput
-
-## PublicClient L1
-
-The following methods are provided to interact with the Cartesi Rollups Smart Contracts.
-
-- estimateAddInputGas
-- estimateDepositEtherGas
-- estimateDepositERC20TokensGas
-- estimateDepositERC721TokenGas
-- estimateDepositSingleERC1155TokenGas
-- estimateDepositBatchERC1155TokenGas
-- estimateExecuteOutputGas
-- validateOutput
+This extension provides L2 actions to interact with the Cartesi Rollups Node, and typed ABIs and addresses of the Cartesi Rollups smart contracts for L1 operations.
 
 ## PublicClient L2
 
@@ -45,12 +19,32 @@ The following methods are provided to interact with the Cartesi Rollups Node.
 - getOutput
 - getReport
 - getProcessedInputCount
-- getLastAcceptedEpoch
+- getLastAcceptedEpochIndex
 - waitForInput
+
+## L1 operations
+
+Transactions related to Cartesi applications are sent directly to the base layer (L1).
+Those are regular contract interactions, best served by viem's own [contract actions](https://viem.sh/docs/contract/writeContract) combined with the typed ABIs and deployment addresses exported by `@cartesi/viem/abi`:
+
+```typescript
+import { inputBoxAbi, inputBoxAddress } from "@cartesi/viem/abi";
+
+const hash = await walletClient.writeContract({
+    abi: inputBoxAbi,
+    address: inputBoxAddress,
+    functionName: "addInput",
+    args: [application, payload],
+});
+```
+
+Users of [wagmi](https://wagmi.sh) can use the generated actions and React hooks from `@cartesi/wagmi`, or generate their own code using the `@cartesi/wagmi-plugin` plugin for `@wagmi/cli`.
 
 ## Utilities
 
-- getInputsAdded: this provides a utility method to get the input(s) added to the InputBox given a `TransactionReceipt`.
+- getInputsAdded: get the input(s) added to the InputBox given a `TransactionReceipt`.
+- getOutputsExecuted: get the output(s) executed by an application given a `TransactionReceipt`.
+- toEVM: convert an `Output` returned by the node API into the arguments of `IApplication.executeOutput` / `IApplication.validateOutput`.
 
 ```typescript
 const receipt = await publicClient.waitForTransactionReceipt({ hash });
@@ -59,13 +53,14 @@ const [inputAdded] = getInputsAdded(receipt);
 
 ## Example
 
-Find below a complete example of depositing ERC-20 tokens and waiting for its effect on L2.
+Find below a complete example of sending an input and waiting for its effect on L2.
 
 ```typescript
-import { createPublicClient, createWalletClient, http, parseUnits } from "viem";
+import { createCartesiPublicClient, getInputsAdded } from "@cartesi/viem";
+import { inputBoxAbi, inputBoxAddress } from "@cartesi/viem/abi";
+import { createPublicClient, createWalletClient, http, stringToHex } from "viem";
 import { privateKeyToAccount } from "viem/accounts";
 import { foundry } from "viem/chains";
-import { getInputsAdded, publicActionsL2, walletActionsL1 } from "../src";
 
 async function main() {
     const chain = foundry;
@@ -76,29 +71,23 @@ async function main() {
     // create public client to chain default url
     const publicClient = createPublicClient({ chain, transport: http() });
 
-    // create extended public client to L2 with RPC url
+    // create cartesi public client to L2 with RPC url
     const publicClientL2 = createCartesiPublicClient({
         transport: http("http://127.0.0.1:6751/rpc"),
     });
 
-    // create extended wallet client to chain default url
-    const walletClient = createWalletClient({
-        chain,
-        transport: http(),
-    }).extend(walletActionsL1());
+    // create wallet client to chain default url
+    const walletClient = createWalletClient({ account, chain, transport: http() });
 
     // application address
     const application = "0xab7528bb862fb57e8a2bcd567a2e929a0be56a5e";
-    const token = "0x92C6bcA388E99d6B304f1Af3c3Cd749Ff0b591e2";
 
     // send input transaction
-    const hash = await walletClient.depositERC20Tokens({
-        account,
-        application,
-        chain,
-        token,
-        amount: parseUnits("1", 18),
-        execLayerData: "0x",
+    const hash = await walletClient.writeContract({
+        abi: inputBoxAbi,
+        address: inputBoxAddress,
+        functionName: "addInput",
+        args: [application, stringToHex("hello")],
     });
 
     // wait for receipt
@@ -107,21 +96,14 @@ async function main() {
     // get input index from receipt
     const [inputAdded] = getInputsAdded(receipt);
     if (inputAdded) {
-        const { inputIndex } = inputAdded;
+        const { index: inputIndex } = inputAdded;
 
         // wait for input to be processed
         const input = await publicClientL2.waitForInput({
-            inputNumber: Number(inputIndex),
+            application,
+            inputIndex,
         });
         console.log(input);
-
-        // get notice 0 produced by application
-        const notice = await publicClientL2.getNotice({
-            inputNumber: Number(inputIndex),
-            noticeNumber: 0,
-        });
-
-        console.log(notice);
     }
 }
