--- packages/viem/src/actions/estimateDepositEtherGas.ts
@@ -1,92 +1,0 @@
-import type {
-    Account,
-    Address,
-    Chain,
-    Client,
-    DeriveChain,
-    EstimateContractGasErrorType,
-    EstimateContractGasParameters,
-    FormattedTransactionRequest,
-    GetChainParameter,
-    Hex,
-    Transport,
-    UnionOmit,
-} from "viem";
-import { estimateContractGas } from "viem/actions";
-import { etherPortalAbi, etherPortalAddress } from "../rollups.js";
-import type {
-    ErrorType,
-    GetAccountParameter,
-    UnionEvaluate,
-} from "../types/utils.js";
-
-export type EstimateDepositEtherGasParameters<
-    chain extends Chain | undefined = Chain | undefined,
-    account extends Account | undefined = Account | undefined,
-    chainOverride extends Chain | undefined = Chain | undefined,
-    _derivedChain extends Chain | undefined = DeriveChain<chain, chainOverride>,
-> = UnionEvaluate<
-    UnionOmit<
-        FormattedTransactionRequest<_derivedChain>,
-        "accessList" | "data" | "from" | "gas" | "gasPrice" | "to" | "type"
-    >
-> &
-    GetAccountParameter<account, Account | Address> &
-    GetChainParameter<chain, chainOverride> & {
-        /** Gas limit for transaction execution */
-        gas?: bigint | undefined;
-        application: Address;
-        execLayerData: Hex;
-    };
-export type EstimateDepositEtherGasReturnType = bigint;
-export type EstimateDepositEtherGasErrorType =
-    | EstimateContractGasErrorType
-    | ErrorType;
-
-export const estimateDepositEtherGas = <
-    transport extends Transport,
-    chain extends Chain | undefined,
-    account extends Account | undefined,
-    chainOverride extends Chain | undefined = undefined,
->(
-    client: Client<transport, chain, account>,
-    parameters: EstimateDepositEtherGasParameters<
-        chain,
-        account,
-        chainOverride
-    >,
-) => {
-    const {
-        account,
-        chain = client.chain,
-        application,
-        gas,
-        maxFeePerGas,
-        maxPriorityFeePerGas,
-        nonce,
-        execLayerData,
-        value,
-    } = parameters;
-
-    const params = {
-        account,
-        abi: etherPortalAbi,
-        address: etherPortalAddress,
-        functionName: "depositEther",
-        args: [application, execLayerData],
-        gas,
-        maxFeePerGas,
-        maxPriorityFeePerGas,
-        nonce,
-        value,
-        // TODO: Not sure `chain` is necessary since it's not used downstream
-        // in `estimateContractGas` or `estimateGas`
-        // @ts-ignore
-        chain,
-    } satisfies EstimateContractGasParameters<
-        typeof etherPortalAbi,
-        "depositEther"
-    >;
-    // biome-ignore lint/suspicious/noExplicitAny: viem pattern
-    return estimateContractGas(client, params as any);
-};
