+++ packages/viem/src/actions/index.ts
@@ -1,73 +1,3 @@
-export {
-    addInput,
-    type AddInputParameters,
-    type AddInputReturnType,
-} from "./addInput.js";
-export {
-    depositBatchERC1155Token,
-    type DepositBatchERC1155TokenParameters,
-    type DepositBatchERC1155TokenReturnType,
-} from "./depositBatchERC1155Token.js";
-export {
-    depositERC20Tokens,
-    type DepositERC20TokensParameters,
-    type DepositERC20TokensReturnType,
-} from "./depositERC20Tokens.js";
-export {
-    depositERC721Token,
-    type DepositERC721TokenParameters,
-    type DepositERC721TokenReturnType,
-} from "./depositERC721Token.js";
-export {
-    depositEther,
-    type DepositEtherParameters,
-    type DepositEtherReturnType,
-} from "./depositEther.js";
-export {
-    depositSingleERC1155Token,
-    type DepositSingleERC1155TokenParameters,
-    type DepositSingleERC1155TokenReturnType,
-} from "./depositSingleERC1155Token.js";
-export {
-    estimateAddInputGas,
-    type EstimateAddInputGasParameters,
-    type EstimateAddInputGasReturnType,
-} from "./estimateAddInputGas.js";
-export {
-    estimateDepositBatchERC1155TokenGas,
-    type EstimateDepositBatchERC1155TokenGasParameters,
-    type EstimateDepositBatchERC1155TokenGasReturnType,
-} from "./estimateDepositBatchERC1155TokenGas.js";
-export {
-    estimateDepositERC20TokensGas,
-    type EstimateDepositERC20TokensGasParameters,
-    type EstimateDepositERC20TokensGasReturnType,
-} from "./estimateDepositERC20TokensGas.js";
-export {
-    estimateDepositERC721TokenGas,
-    type EstimateDepositERC721TokenGasParameters,
-    type EstimateDepositERC721TokenGasReturnType,
-} from "./estimateDepositERC721TokenGas.js";
-export {
-    estimateDepositEtherGas,
-    type EstimateDepositEtherGasParameters,
-    type EstimateDepositEtherGasReturnType,
-} from "./estimateDepositEtherGas.js";
-export {
-    estimateDepositSingleERC1155TokenGas,
-    type EstimateDepositSingleERC1155TokenGasParameters,
-    type EstimateDepositSingleERC1155TokenGasReturnType,
-} from "./estimateDepositSingleERC1155TokenGas.js";
-export {
-    estimateExecuteOutputGas,
-    type EstimateExecuteOutputGasParameters,
-    type EstimateExecuteOutputGasReturnType,
-} from "./estimateExecuteOutputGas.js";
-export {
-    executeOutput,
-    type ExecuteOutputParameters,
-    type ExecuteOutputReturnType,
-} from "./executeOutput.js";
 export { getApplication } from "./getApplication.js";
 export { getChainId } from "./getChainId.js";
 export { getCommitment } from "./getCommitment.js";
@@ -92,9 +22,4 @@ export { listOutputs } from "./listOutputs.js";
 export { listReports } from "./listReports.js";
 export { listTournaments } from "./listTournaments.js";
 export { listWithdrawals } from "./listWithdrawals.js";
-export {
-    validateOutput,
-    type ValidateOutputParameters,
-    type ValidateOutputReturnType,
-} from "./validateOutput.js";
 export { waitForInput } from "./waitForInput.js";
