+++ packages/viem/src/index.ts
@@ -3,8 +3,7 @@ export {
     createCartesiPublicClient,
     type CartesiPublicClient,
 } from "./clients/createCartesiPublicClient.js";
-export { publicActionsL1 } from "./decorators/publicL1.js";
 export { publicActionsL2 } from "./decorators/publicL2.js";
-export { walletActionsL1 } from "./decorators/walletL1.js";
 export * from "./types/actions.js";
+export { type ExecuteOutputArgs, toEVM } from "./types/output.js";
 export * from "./utils/index.js";
